requirements.txt
crewai==0.201.1
faiss_cpu==1.12.0
fastapi==0.118.0
fastmcp==2.12.4
httpx==0.28.1
langchain==0.3.27
langchain_community==0.3.31
pandas==2.3.3
pdfplumber==0.11.7
pydantic==2.12.0
pydantic_settings==2.11.0
PyJWT==2.10.1
PyPDF2==3.0.1
sentence_transformers==5.1.1
streamlit==1.50.0
uvicorn==0.37.0
langsmith=0.4.31
openinference-instrumentation-crewai=0.1.13
